requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
Pillow==9.3.0
psycopg2==2.9.5
pytz==2022.6
sqlparse==0.4.3
